fastApi/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pymongo==4.3.3
pytest==7.3.1
pytest-cov==4.1.0
httpx==0.24.1
black==23.3.0
flake8==6.0.0